requirements.txt
diffusers==0.25.1
huggingface-hub==0.25.2
einops==0.8.0
imageio==2.36.1
insightface==0.7.3
matplotlib==3.10.0
numpy
opencv_python==4.10.0.84
packaging==24.2
Pillow==11.1.0
scipy==1.15.0
torch==2.1.1+cu118
torchvision==0.16.1+cu118
tqdm==4.67.1
transformers==4.42.4
